pyproject.toml
[project]
name = "team-03-2-session03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "specklepy>=3.2.4",
]
